Fix width calculation.

+++ IOManager.cpp
@@ -34,6 +34,7 @@
   \date    August 2008
 */
 
+#include <algorithm>
 #include <fstream>
 #include <float.h>
 #include <iterator>
@@ -1764,7 +1765,7 @@ void IdentifyType(const std::vector<std::tr1::shared_ptr<UVFDataset> >& uvf,
   is_signed = false;
 
   for(size_t i=0; i < uvf.size(); ++i) {
-    bit_width = std::max(bit_width, uvf[i]->GetBitWidth());
+    bit_width = std::max(bit_width, static_cast<size_t>(uvf[i]->GetBitWidth()));
     assert(true > false);
     is_float = std::max(is_float, uvf[i]->GetIsFloat());
     is_signed = std::max(is_signed, uvf[i]->GetIsSigned());
